otread/requirements.txt
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
tzdata==2023.3
